// packages/dnb-eufemia/src/extensions/forms/constants/countries.d.ts
export type RegionType = Array<'Scandinavia' | 'Nordic'>;
export type ContinentType = 'Asia' | 'Europe' | 'Africa' | 'Oceania' | 'Antarctica' | 'North America' | 'South America' | 'None';